@@ -1,6 +1,6 @@
 {
   "name": "@sensenet/list-controls-react",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "",
   "main": "dist/ContentList/index.js",
   "files": [
@@ -53,39 +53,38 @@
   },
   "homepage": "https://sensenet.com",
   "devDependencies": {
-    "@types/chai": "^4.1.4",
+    "@types/chai": "^4.1.5",
     "@types/mocha": "^5.2.5",
-    "@types/react": "^16.4.9",
+    "@types/react": "^16.4.14",
     "@types/react-dom": "^16.0.7",
-    "@types/react-router": "^4.0.30",
-    "@types/react-router-dom": "^4.3.0",
+    "@types/react-router": "^4.0.31",
+    "@types/react-router-dom": "^4.3.1",
     "@types/react-test-renderer": "^16.0.2",
-    "@types/uuid": "^3.4.3",
-    "awesome-typescript-loader": "^5.2.0",
+    "@types/uuid": "^3.4.4",
+    "awesome-typescript-loader": "^5.2.1",
     "chai": "^4.1.2",
     "copy-webpack-plugin": "^4.5.2",
     "mocha": "^5.2.0",
-    "nyc": "^12.0.2",
+    "nyc": "^13.0.1",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "source-map-loader": "^0.2.3",
+    "source-map-loader": "^0.2.4",
     "tslint": "^5.11.0",
-    "typescript": "^3.0.1",
-    "webpack": "^4.16.5",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.5"
+    "typescript": "^3.0.3",
+    "webpack": "^4.19.1",
+    "webpack-bundle-analyzer": "^3.0.2",
+    "webpack-cli": "^3.1.1",
+    "webpack-dev-server": "^3.1.8"
   },
   "dependencies": {
-    "@material-ui/core": "^1.5.0",
-    "@material-ui/icons": "^1.1.0",
+    "@material-ui/core": "^3.1.1",
+    "@material-ui/icons": "^3.0.1",
     "@sensenet/default-content-types": "^1.1.0",
     "moment": "^2.22.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-fontawesome": "^1.6.1",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
     "react-moment": "^0.7.9",
-    "react-test-renderer": "^16.4.2",
+    "react-test-renderer": "^16.5.2",
     "uuid": "^3.3.2"
   },
   "config": {
