@@ -1,4 +1,4 @@
-import { TableCell } from '@material-ui/core'
+import TableCell from '@material-ui/core/TableCell'
 import * as React from 'react'
 import Moment from 'react-moment'
 
