@@ -1,5 +1,17 @@
-import { AppBar, Drawer, IconButton, List, ListItem, ListItemIcon, ListItemText, Toolbar, Typography } from '@material-ui/core'
-import { Home, List as ListIcon, Menu } from '@material-ui/icons'
+import AppBar from '@material-ui/core/AppBar'
+import Drawer from '@material-ui/core/Drawer'
+import IconButton from '@material-ui/core/IconButton'
+import List from '@material-ui/core/List'
+import ListItem from '@material-ui/core/ListItem'
+import ListItemIcon from '@material-ui/core/ListItemIcon'
+import ListItemText from '@material-ui/core/ListItemText'
+import Toolbar from '@material-ui/core/Toolbar'
+import Typography from '@material-ui/core/Typography'
+
+import Home from '@material-ui/icons/Home'
+import ListIcon from '@material-ui/icons/List'
+import Menu from '@material-ui/icons/Menu'
+
 import * as React from 'react'
 import { Route, Switch } from 'react-router'
 import { HashRouter, Link } from 'react-router-dom'
